Fix waitForExits blocking when flush channel is full
When a disconnect occurs, the flusher will exit but there may be still multiple calls to kickFlusher() that would cause the flush channel to be full (created buffered with length 1024). When that occurs and spinUpSocketWatchers starts and calls waitForExits, the call to nc.fch <- true would block because the channel is full and nothing is empting it. That would cause the reconnect process to "fail" blocking in spinUpSocketWatchers.