Changing test's Wait(chan bool) timeout value
Some tests are randomly failing on Travis, and I suspect that this is a timing issue. Try to solve this by bumping the timeout from 500 ms to 5 seconds. That is important to know for tests that rely on Wait() to timeout after a short amount of time. I had to change 2 such tests.