Remove sub.mch as a parameter to deliverMsgs() goroutine since we now pass sub.
Since we pass 'sub' as a parameter to deliverMsgs, there is no need to pass sub.mch. We are getting it from the 'sub' at the beginning of the goroutine. Note that we need to check if the channel is not nil before starting. There is a possibility that we get unsubscribed right after the subscription is created.