Fixed reconnect loop not yielding to Close()
On Windows, creating a TCP connection takes more than 1 second
when no server is running at the given address. If the reconnect
wait is lower, it will cause the doReconnect routine to loop
through the reconnect attempts without ever releasing the connection
lock, which prevents a parallel Close() from breaking out the loop.