Fix Chan subscribers inability to receive more than 65536 msgs
All variants of Chan subscribers would stop receiving messages
after more than 65536 messages or 65536*1024 bytes. This was
because on processing of a message, the pending counts were
incrememted but never decremented, which would lead to further
messages to be dropped. We no longer update the pending counts
for ChanSubscriptions since they are not relevant.

Resolves #193