[FIXED] Possible panic on WaitGroup.Wait()
There have been 2 reports in the last 10 months of this happening
once each. We have not been able to reproduce the issue. However,
this code change should eliminate the possibility of code going
into a Wait() while in parallel doing an Add() (for the same group).
There is no new test since we have multiple tests that check for
reconnect and we have not been able to reproduce. A long running
test causing 100,000 server restarts at randome interval has been
running on *original code* without producing the panic.

Attempt to resolve #192