[ADDED] Handler invoked when a new server joins the NATS Cluster
If DiscoveredServersCB is set, the client library will invoke this
callback when it is notified by the server it connects to that
a new server has joined the cluster. This callback is not invoked
in the initial Connect(). A ConnHandler is used. The user can
invoke nc.DiscoveredServers() to get the list of discovered servers.

But in some cases, it is just enough to know that a new server
has joined. A use case is in NATS Streaming server with the
channels partioning/sharding feature where a server needs to know
if the cluster has changed in which case it should resend its list
of channels.