[ADDED] Conn.Barrier() API
Suppose that you have a connection with two subscriptions, let's
call them "pub" and "close".
Let's say that you know that when receiving a message on "close",
no new message will arrive on "pub" (because the sender does
send "pub" and then a final message on "close"). But because
"pub" and "close" are two different subscriptions, there is no
guarantee that the message on "close" is not going to be dispatched
before the last message on "pub".

The Barrier() API can be used to solve this issue. In the example
above, if the message callback on "close" were to call that API,
this will ensure that all pending messages on "pub" have been
dipatched before the provided function in the Barrier() API gets
executed.

See `test/TestBarrier()` test for more details.