Check for EOF and update some tests due to changes in gnatsd tests
In some cases the server may send an authorization violation and
closes the connection right away. With the original change, the
test in server would fail. Handling the EOF here solves the issue.
From tests it looks like the last `\n` is often missing but it
is fine since the client library when not getting a PONG assumes
it is an error and strips away the `-ERR` prefix and the new line
suffix anyway.

Updated a test since we changed the server DefaultTestOptions
port to be -1 instead of 4222.