Check if context canceled before waiting for Msg
Context can be used to cancel blocking waiting for a message, and in
the case that the context has already been canceled it is expected for
it to return right away since the context no longer valid.

Previous behavior was not deterministic since it was possible for a
client to still receive a messages even though the context was already
canceled.  In order to fix this, now we check first whether the
context has been canceled previous to waiting for the message to
prevent successfully making requests with invalid contexts.

Signed-off-by: Waldemar Quevedo <wally@synadia.com>