[CHANGED] DefaultURL changed to "nats://127.0.0.1:4222"
Replaced to loopback ipv4 instead of localhost. In some cases,
when running the server with 127.0.0.1, if the application uses
the previous default URL (localhost), the library would resolve
this and possibly get 2 IPs (common to get 127.0.0.1 and ::1),
and it would then pick randomly one of those 2. If getting ::1,
and the server is listening on 127.0.0.1, the connection would
fail before it tries with the ipv4 address.

Also replaced reference of gnatsd with nats-server.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>