Make code examples more copy-paste friendly
Currently, there are some code examples that omit the package name. This requires the user to do a little more work on their end. They would have to know that they need the package name.

This change adds the package name to the code so it's easier to copy-paste examples.