[FIXED] Request() with UseOldRequest not kicked out on Close()
Applies only when UseOldRequest() options is passed to Connect():

If a Request() was issued and the connection closed, the Request()
call would not return with "connection closed" error, instead it
would wait for the provided timeout.

The issue would not exist in the RequestWithContext() API, even
when using the old style request, but a test has been added
nevertheless.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>