Make Msg.Header fields case-preserving when accessed directly
There is a common issue with Go and its HTTP libraries in that responses handled
by its HTTP clients become represented in 'canonical form' (`X-Foo-Bar`),
sometimes causing integration issues with proxies, some APIs, legacy
apps and likely compatibility with http/2 headers to some degree which
requires headers in lowercase, (unlike http v1 ones that are case insensitive).

As a workaround, the Go docs suggest to access the Header map directly
to use the fields not in canonical form.

In this PR, it is avoided during the decoding of the headers calling
`CanonicalMIMEHeaderKey` in order to be able to preserve the raw version
of the headers when traversing over multiple hops without causing
unexpected side effects or changing the format over the wire.

If users add headers via Header.Add() or Header.Set() for example, those
methods already call `CanonicalMIMEHeaderKey` when creating the header
so they will be presented in canonical form without changes.

Signed-off-by: Waldemar Quevedo <wally@synadia.com>