CreateKeyValue with RePublish will fail if KV already exists
This cannot be updated in the server at the moment, so if user
calls CreateKeyValue() with a config that does not have RePublish
and then call again on the same bucket with RePublish set, the
user will get an error. Added a test that shows that this is the case.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>