Fix TestJetStreamStreamInfoWithSubjectDetails test
Since 2.9.15, the server handles inbound messages in a separate execution context, which is an optimisation, but it also means clients can't rely on flushing published messages to ensure that JetStream has it stored and accounted. That behaviour was never guaranteed, and it worked only because of the lack of this optimisation.

Signed-off-by: Tomasz Pietrek <tomasz@nats.io>