Fixed issue with writing compressed data
With compress/flate, it seems that after calling Close() it was
fine dropping the last 4 bytes before doing the framing, but
with the different flate, it fails.

Instead, use Flush() (like in the server) and then dropping the
last 4 bytes works as before.

I have added a compression.Close() on connection close, after
sending the CLOSE frame.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>