Proposal to provide config based deactivation of QueueGroups for Service.
Not every service is required to be in a QueueGroup.
Some services require direct address and the code changes presented here are for discussion.

Is this an acceptable way to turn off QueueGroups? It works in my testing, but was wondering if there are better suggestions?

Naturally, these changes have dependencies. The Schema entry for the services info_response will need changing, to allow for the tokensied naming with a period.

This definition https://nats.io/schemas/micro/v1/info_response.json specifically requires `basic_name` to allow the period and it is proposed to create a `micro_name` definition that includes the period.

Just to be clear, I appreciate that this might not be the solution, but as we do more with Service, this has come up and needs to be addressed. I'm happy to make the various PRs to extend this, but wanted to seek opinion and counsel before blindly going after a PR!

If you're curious, you can export the env var `NOVALIDATE="nope"` for the `nats` CLI client and stand up some period tokenised services and get info using this PR.

```go
	group := srv.AddGroup("test") // the subject token pattern here is "test.get.1234"
	err = group.AddEndpoint("get.1234", micro.HandlerFunc(getHandler), micro.WithDisableQueueGroup(true))
```

Cheers,
Dave

Signed-off-by: davedotdev <me@dave.dev>